conclusao do back-end, faltando alguns ajustes

+++ src/main/java/br/com/banco/dtos/TransferenciaContaDTO.java
@@ -1,6 +1,7 @@
 package br.com.banco.dtos;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.Date;
 
 import lombok.AllArgsConstructor;
@@ -13,7 +14,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class TransferenciaContaDTO {
-    private Long idConta;
+    private Integer idConta;
     private String nomeResponsavel;
     private Long idTransferencia;
     private Date dataTransferencia;
